Fixed Kel's Bag O Words summary printing.
After my change to using optparse, Kel's Bag O
Words summaries were always printing with a
percetnage of 0%. This was because I was dividing
an int by and int, with the top always less than
the bottom (whoops). Changed to a floating point.

Also, a directory for collecting sample data
from Twitter. Included a sample during a PRT
outage on December 16th, 2010.